Started a new system to handle jumps better.
Currently I have an immediate-mode style system. I need to add a system like an
assembler stage in order to calculate the number of bytes the jumps need to jump
over. In order to accomplish this I will also be writing an allocator. Each
subsystem really should be using an allocator instead of an array and I'd rather
just do it now than have the next stage also use the array in the main function.

I had a lot more written out before with more of the reasoning behind this and
the details. I accidentally crashed my computer by filling up the RAM though and
lost the message. You'll have to see the details by reading the code.