Finished up the memory code refactoring I started a few commits ago
Honestly I'm not sure this is really simpler than it was before. The error
returns and bug fixes are nice. It's just that I ended up adding 3 more
functions to the platform API. That doesn't seem simpler. I was really hoping
that Expand could be platform agnostic.

I guess the windows version is sort of the platform agnostic version, but we
don't have any other platforms anyway.

- ArenaInit is platform agnostic now
- Added ArenaFree
- Cleaned up the platform Expand behavior.
- Moved platform code into one file
- Arena functions return errors properly instead of Asserting
- Fixed some bugs in Expand behavior