Completed the x86_opcode tests for all non-jump instructions
 - Added some asserts in the assembler to enforce 0's where values are ignored
 - Fixed a bug with the tester state not getting 0-initialized

I used this string of vim regexes to convert the output strings for missing test
data to the syntax of the want data:

:'<,'>s/\([0-9A-F]\{2}\) /0x\1, /g | '<,'>s/, $// | '<,'>s/\(\([0-9A-Fx]\{4}\(, \)\=\)\+\)/WantOp(\1)/

Note: it works when you highlight the lines you want to act on with V

I pasted the output code into https://defuse.ca/online-x86-assembler.htm
to manually check that the output was correct.

It would be possible to output all the testcases in one big file you could run
through ndisasm to manually re-check it.