Add an end-to-end tests file
 - Add const qualifiers where needed in a few places since I need const
   for the string literals
 - Add a compiler flag to not put in mutex stuff for static variables in
   functions
 - Wrote a self-contained CompileRegex function that we might want to
   just make a library call we support. Not sure about the static
   arenas...
 - Fixed compiling the tester after I forgot to delete the old includes
   I removed in the commit before this

I actually wrote CompileRegex and all of the test data a while back (who
knows when), but I was trying to do it by making an array of test case
structs which needed two inferred-size arrays each for the match and
no-match data. This syntax doesn't really work unless you're willing to
use std::initializer_list or an existing class that uses it (like
std::vector).

Today I had some time and pulled out this project and figured out that
this syntax issue was why I never committed this. After a few tries and
some thinking I came up with a better option to put in this data, then
had to fix several compiler errors.