Support compiling regex with arbitrarily many NFA states!
Previously the limit was only 32 because we use a single register, now
it uses the stack for the bit storage.

Another interesting change is I'm also using the arena storage for an
auxilary buffer needed to calculate the ActivateMask (which was
previously a uint32_t). I can use this same idea to store the buffer of
parentheses bounds in the parser to support arbitrarily many parens
(with an extra pass over the input string to count the parens).

I also randomly found a bug with the NFA generation for regex like
'a+|b', which will match ab in the current code but it shouldn't. The
issue is the a+ causes an epsilon arc to loop back to the beginning of
the alternative choice which allows the b path to be taken. Instead I
need to create a extra state for alternative clauses, which means I need
to be able to change the start state. So, the tests fail in this commit.