Fixed a bug in NFA generation for alternatives Also fix not reporting an error for adding a loop char when not allowed
Previously regex like a+|b would fail because the + would loop back to
the start of the alternative check instead of just the start of the a
clause. So we needed an extra start state.

For top-level alternatives we replace the start state with a new one
only when we see the first | char, so we avoid adding a state to all
NFAs that we generate. Unfortunately I could not do the same for
parentheses groups, I had to add a second new state. I believe the first
new state for parentheses groups was actually added for alternatives
too.

At least we support arbitrarily many states now so it's not so bad to
add states. Maybe someday we'll have an NFA optimization pass at the
end.